+++ src/components/List/CoverList.vue
@@ -205,7 +205,7 @@ const playList = debounce(
       const list = await getListData(item.id);
       player.updatePlayList(list, undefined, item.id as number);
     } catch (error) {
-      console.log("Error to play: ", error);
+      console.error("Error to play: ", error);
     } finally {
       item.loading = false;
     }